+++ src/textField/story.tsx
@@ -8,36 +8,39 @@ import TextField from 'textField'
 import ArrowIcon from 'icon/arrowIcon'
 import Button from 'button'
 import readme from 'textField/specifications/textField.md'
-import primaryDoc from "../button/specifications/primary.md"
+import primaryDoc from '../button/specifications/primary.md'
 
 const stories = storiesOf('TextField', module)
 stories.addDecorator(withKnobs)
 
 const inputTypes = ['text', 'email', 'number', 'password']
 
-stories.add('specifications',
-    () => (<TextField
-        type={select('type', inputTypes, 'text')}
-        id={text('id')}
-        name={text('name', 'inputName')}
-        placeholder={text('placeholder')}
-        labelledBy={text('aria label')}
-        disabled={boolean('disabled', false)}
-        readOnly={boolean('readOnly', false)}
-        label={text('label')}
-        error={text('error message', '')}
-        onChange={action('changed')}
-        onFocus={action('focused')}
-        onBlur={action('blurred')}
-        focus={boolean('focus', false)}
-        required={boolean('required', false)}
-        maxLength={number('maxLength')}
-        autoComplete={select('autocomplete', ['on', 'off'])}
-        title={text('title', 'accessibility text')}
-    />),
-    {
-      readme: { content: readme },
-    },
+stories.add(
+  'specifications',
+  () => (
+    <TextField
+      type={select('type', inputTypes, 'text')}
+      id={text('id')}
+      name={text('name', 'inputName')}
+      placeholder={text('placeholder')}
+      labelledBy={text('aria label')}
+      disabled={boolean('disabled', false)}
+      readOnly={boolean('readOnly', false)}
+      label={text('label')}
+      error={text('error message', '')}
+      onChange={action('changed')}
+      onFocus={action('focused')}
+      onBlur={action('blurred')}
+      focus={boolean('focus', false)}
+      required={boolean('required', false)}
+      maxLength={number('maxLength')}
+      autoComplete={select('autocomplete', ['on', 'off'])}
+      title={text('title', 'accessibility text')}
+    />
+  ),
+  {
+    readme: { content: readme },
+  },
 )
 
 stories.add('input', () => (
@@ -179,7 +182,11 @@ stories.add('textarea', () => (
   />
 ))
 
-const formatValue: (value: string) => string = value => {
+const formatValue = (value: string, previousValue: string) => {
+  if (previousValue && value.length < previousValue.length) {
+    // Removing character, do not escape
+    return value
+  }
   if (value.match(/^[0-9]{2}$/) || value.match(/^[0-9]{2}\/[0-9]{2}$/)) {
     return `${value}/`
   }
@@ -198,7 +205,7 @@ stories.add('date input ', () => (
     onChange={action('changed')}
     onFocus={action('focused')}
     onBlur={action('blurred')}
-    maxLength={number('maxLength')}
+    maxLength={number('maxLength', 10)}
     autoComplete={select('autocomplete', ['on', 'off'])}
     format={formatValue}
   />